Refactor DbContextBase and remove unnecessary code
This commit refactors the DbContextBase class to simplify and optimize it. It removes redundant code and improves effectiveness of SaveChanges methods by isolating SaveCreated and SaveUpdated functions. Additionally, unnecessary code in OnModelCreating method was removed to improve the performance. The changes also include minor formatting alterations for readability and code consistency.